pyproject.toml
[project]
name = "pulsar-asyncio-client"
description = "Wrapper for pulsar-client that publishes messages asynchronously and handles consumed messages in an asyncio event loop"
authors = [
    {name = "Mark Parncutt", email = "me@markparncutt.com"},
]
dynamic = ["version"]
dependencies = [
    "pulsar-client>=3.5.0"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.version]
source = "scm"
write_to = "pulsar_asyncio/version.py"
write_template = "# Dynamically generated by build system, do not modify.\n__version__ = version = \"{}\"\n"

[tool.pdm.build]
excludes = ["./**/.git"]
includes = ["pulsar_asyncio/"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
test = ["pytest>=7.4.3"]
